requirements.txt
pytelegrambotapi~=4.29.1
psycopg2-binary~=2.9.11
openai~=2.9.0
python-dotenv~=1.0.1
